Finish the exercise to specification
In this exercise I ran into bug that I couldn't fix
but the solution suffered from the same bug. I wanted
the cards to be a maximum of 320px in height, to shrink
when the window is resized, to wrap on a row, and for
that row to be near the previous row. Could not achieve
this and defaulted to my original solution that works
like the given solution.